+++ app/Client/Mapper/AbstractFilterMapper.php
@@ -64,7 +64,7 @@ abstract class AbstractFilterMapper
 
     private static function isAssocAndNotEmpty(array $arr): bool
     {
-        if ([] === $arr) {
+        if (empty($arr)) {
             return false;
         }