+++ app/Repository/CampaignRepository.php
@@ -82,6 +82,9 @@ class CampaignRepository
 
         try {
             $campaign->delete();
+            foreach ($campaign->banners as $banner) {
+                $banner->classifications()->delete();
+            }
             $campaign->banners()->delete();
         } catch (\Exception $ex) {
             DB::rollBack();
@@ -116,6 +119,7 @@ class CampaignRepository
 
             if ($bannersToDelete) {
                 foreach ($bannersToDelete as $banner) {
+                    $banner->classifications()->delete();
                     $banner->delete();
                 }
             }