+++ src/Classify/Infrastructure/Service/DummyClassifier.php
@@ -88,15 +88,12 @@ class DummyClassifier implements ClassifierInterface
         }
 
         $dummy = $this->banners[$bannerId];
-        $classification = Classification::createUnsigned(
+        $classification = new Classification(
             $this->keyword,
             $dummy['publisherId'],
             $dummy['status'],
             $dummy['siteId']
         );
-
-        $signature = $this->signatureVerifier->create($classification->keyword(), $dummy['bannerId']);
-        $classification->sign($signature);
 
         return new ClassificationCollection($classification);
     }